refactor: 적기 함수 이름 변경
